This project aims to create a practical Android application serving as a digital tool for calculating the volumes of common geometric shapes.
This project is a mobile application developed using Android Studio, designed to be a practical and user-friendly tool for calculating the area and volume of various geometric shapes. The app is ideal for students, engineers, architects, and anyone who needs quick and accurate dimensional calculations without the need for a physical calculator or complex software.